Move GenerateBlock to miner_ext.go
